# cmd/api/docs/swagger.yaml
definitions:
...
  github_com_waqasmani_go-boilerplate_internal_shared_utils.ErrorResponse:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
      type:
        type: string
    type: object
  github_com_waqasmani_go-boilerplate_internal_shared_utils.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.ErrorResponse'
      success:
        type: boolean
      version:
        type: string
    type: object
...
  internal_modules_users.CreateUserRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 8
        type: string
      role:
        enum:
        - user
        - admin
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - role
    type: object
  internal_modules_users.ListUsersResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      users:
        items:
          $ref: '#/definitions/internal_modules_users.UserResponse'
        type: array
    type: object
  internal_modules_users.UpdatePasswordRequest:
    properties:
      new_password:
        minLength: 8
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
  internal_modules_users.UpdateUserRequest:
    properties:
      first_name:
        type: string
      last_name:
        type: string
    required:
    - first_name
    - last_name
    type: object
  internal_modules_users.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      last_name:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
  description: This is an attendance and employee management server.
  title: Go Boilerplate API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Get paginated list of users (admin only)
      parameters:
      - default: 20
        description: Page size
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_users.ListUsersResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
      security:
      - Bearer: []
      summary: List users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new user (admin only)
      parameters:
      - description: User creation details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_users.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_users.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
      security:
      - Bearer: []
      summary: Create user
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a user (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
      security:
      - Bearer: []
      summary: Delete user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Get user details by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_users.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
      security:
      - Bearer: []
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update user's first and last name
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated user details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_users.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_users.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
      security:
      - Bearer: []
      summary: Update user profile
      tags:
      - Users
  /users/{id}/password:
    put:
      consumes:
      - application/json
      description: Change user's password
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Password update details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_users.UpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
      security:
      - Bearer: []
      summary: Update user password
      tags:
      - Users
securityDefinitions:
  Bearer:
    description: Type "Bearer <your_token>" (include the word Bearer and a space)
    in: header
    name: Authorization
    type: apiKey
  CsrfToken:
    description: Enter your CSRF token
    in: header
    name: X-CSRF-TOKEN
    type: apiKey
swagger: "2.0"